refactor(article,draft): integrate draft webmvc adapter to article webmvc adapter
