+++ services/article/driving/web-mvc/src/main/java/nettee/blole/article/web/DraftCommandApi.java
@@ -1,21 +1,21 @@
-package nettee.draft.driving.web;
+package nettee.blole.article.web;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import nettee.draft.application.usecase.DraftCreateUseCase;
-import nettee.draft.application.usecase.DraftDeleteUseCase;
-import nettee.draft.application.usecase.DraftImageCreateUseCase;
-import nettee.draft.application.usecase.DraftUpdateUseCase;
-import nettee.draft.domain.type.DraftStatus;
-import nettee.draft.driving.web.dto.DraftCommandDto.DraftCommandResponse;
-import nettee.draft.driving.web.dto.DraftCommandDto.DraftCreateCommand;
-import nettee.draft.driving.web.dto.DraftCommandDto.DraftImageCreateResponse;
-import nettee.draft.driving.web.dto.DraftCommandDto.DraftUpdateCommand;
-import nettee.draft.driving.web.mapper.DraftDtoMapper;
+import nettee.blolet.article.domain.sub.DraftStatus;
 import nettee.blolet.jwt.filter.annotation.AuthUser;
 import nettee.blolet.jwt.filter.annotation.AuthorizedUser;
+import nettee.blolet.article.application.usecase.DraftCreateUseCase;
+import nettee.blolet.article.application.usecase.DraftDeleteUseCase;
+import nettee.blolet.article.application.usecase.DraftImageCreateUseCase;
+import nettee.blolet.article.application.usecase.DraftUpdateUseCase;
+import nettee.blole.article.web.dto.DraftCommandDto.DraftCommandResponse;
+import nettee.blole.article.web.dto.DraftCommandDto.DraftCreateCommand;
+import nettee.blole.article.web.dto.DraftCommandDto.DraftImageCreateResponse;
+import nettee.blole.article.web.dto.DraftCommandDto.DraftUpdateCommand;
+import nettee.blole.article.web.mapper.DraftDtoMapper;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
