+++ services/article/driving/web-mvc/src/main/java/nettee/blole/article/web/mapper/DraftDtoMapper.java
@@ -1,0 +1,23 @@
+package nettee.blole.article.web.mapper;
+
+import nettee.blole.article.web.dto.DraftCommandDto.DraftCreateCommand;
+import nettee.blole.article.web.dto.DraftCommandDto.DraftImageCreateResponse;
+import nettee.blole.article.web.dto.DraftCommandDto.DraftUpdateCommand;
+import nettee.blole.article.web.dto.DraftQueryDto.DraftDetailResponse;
+import nettee.blolet.article.domain.Draft;
+import nettee.blolet.article.domain.DraftImage;
+import nettee.blolet.article.domain.sub.DraftStatus;
+import nettee.blolet.article.readmodel.DraftReadModels.DraftDetail;
+import org.mapstruct.Mapper;
+
+import java.util.Optional;
+
+@Mapper(componentModel = "spring")
+public interface DraftDtoMapper {
+
+    Draft toDomain(DraftCreateCommand command, DraftStatus status);
+    Draft toDomain(String id, DraftUpdateCommand command);
+
+    DraftDetailResponse toDtoDetail(Optional<DraftDetail> board);
+    DraftImageCreateResponse toCreateResponse(DraftImage draftImage);
+}
